Created cleaner abstraction of websocket handling
